Simplifying BinaryIOUtility API (WIP, everything else be broken)
* Remove Decimal, Char, String[] read/write functions
* Remove ability to decide whether or not to create new instance when deserializing IBinarySerializable (calling instance.Deserialize(reader) makes more sense)

+++ Manifold_Unity/Assets/Root/Manifold/IO/Extensions.System.IO.BinaryWriter.Manifold.IO.cs
@@ -191,7 +191,7 @@ namespace Manifold.IO
 
         public static Pointer GetPositionAsPointer(this BinaryWriter writer)
         {
-            return new Pointer((int)writer.BaseStream.Position);
+            return new Pointer(writer.BaseStream.Position);
         }
     }
 }
